requirements.txt
transformers
accelerate
datasets
librosa
chromadb
openai-whisper
termcolor
matplotlib
numpy
pandas
scikit-learn
torchaudio
scipy
tensorboard
moviepy
ffmpeg-python